Fix models cadinality & Quantity field data type.

+++ Models/Product/Product.cs
@@ -1,5 +1,4 @@
 using System.ComponentModel.DataAnnotations;
-using System.ComponentModel.DataAnnotations.Schema;
 using PadariaTech.Models.Base;
 
 namespace PadariaTech.Models
@@ -9,7 +8,7 @@ namespace PadariaTech.Models
         [DataType(DataType.Currency)]
         public decimal Price { get; set; }
 
-        public int StockQuantity { get; set; }
+        public double StockQuantity { get; set; }
 
         [MaxLength(30)]
         public string Measure { get; set; }
